refactor: moved props info to the mocks folder
MovieCard.stories.tsx
MovieCard.spec.tsx
