+++ reactjs-films-homework/src/components/MovieCard/mocks/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MovieCardMock';
